test: run 0 perm tests only for Linux
