+++ pkg/analyze/dir_test.go
@@ -133,32 +133,6 @@ func TestHardlink(t *testing.T) {
 	assert.Equal(t, 'H', dir.Files[0].(*Dir).Files[1].GetFlag())
 }
 
-func TestErr(t *testing.T) {
-	fin := testdir.CreateTestDir()
-	defer fin()
-
-	err := os.Chmod("test_dir/nested", 0)
-	assert.Nil(t, err)
-	defer func() {
-		err = os.Chmod("test_dir/nested", 0755)
-		assert.Nil(t, err)
-	}()
-
-	analyzer := CreateAnalyzer()
-	dir := analyzer.AnalyzeDir(
-		"test_dir", func(_, _ string) bool { return false }, false,
-	).(*Dir)
-	analyzer.GetDone().Wait()
-	dir.UpdateStats(make(fs.HardLinkedItems))
-
-	assert.Equal(t, "test_dir", dir.GetName())
-	assert.Equal(t, 2, dir.ItemCount)
-	assert.Equal(t, '.', dir.GetFlag())
-
-	assert.Equal(t, "nested", dir.Files[0].GetName())
-	assert.Equal(t, '!', dir.Files[0].GetFlag())
-}
-
 func BenchmarkAnalyzeDir(b *testing.B) {
 	fin := testdir.CreateTestDir()
 	defer fin()
