+++ pkg/analyze/file_test.go
@@ -1,7 +1,6 @@
 package analyze
 
 import (
-	"os"
 	"testing"
 	"time"
 
@@ -241,35 +240,6 @@ func TestRemoveFile(t *testing.T) {
 	assert.Equal(t, 1, len(dir.Files))
 	assert.Equal(t, 2, dir.ItemCount)
 	assert.Equal(t, int64(2), dir.Size)
-}
-
-func TestRemoveFileWithErr(t *testing.T) {
-	fin := testdir.CreateTestDir()
-	defer fin()
-
-	err := os.Chmod("test_dir/nested", 0)
-	assert.Nil(t, err)
-	defer func() {
-		err = os.Chmod("test_dir/nested", 0755)
-		assert.Nil(t, err)
-	}()
-
-	dir := &Dir{
-		File: &File{
-			Name: "test_dir",
-		},
-		BasePath: ".",
-	}
-
-	subdir := &Dir{
-		File: &File{
-			Name:   "nested",
-			Parent: dir,
-		},
-	}
-
-	err = RemoveItemFromDir(dir, subdir)
-	assert.Contains(t, err.Error(), "permission denied")
 }
 
 func TestTruncateFile(t *testing.T) {
